Removed Fail implementations
Not totally sure why these were originally needed, but the existing code
was causing warnings due to the introduction of the `dyn` trait, but
changing from `&Fail` to `dyn Fail` gave a compile error due to
`Fail` being a trait and thus unsized.  While that makes sense I did not
have a good idea of the proper way to fix it, so I just removed the need
for it in the first place.